requirements.txt
bioframe
biopython
# bowtie2
click
cooler
cooltools
funcy
hicstuff
matplotlib
numpy
pandas
pysam
# samtools
scipy
sklearn
statsmodels
